brain still fuzzy from puking stupid food-poisoned pi-r-squared pizza all night friday, but committing a bunch of changes to pics anyways. Hopefully not breaking too much here... :P
synopsis: This is the begining of the lastest ground-up revision of the
pics interface.. (given my previous non-tnmc attempts, this is what,
version 5? i wonder if i'll get it right this time.. <sigh> ..) i'm not
modifying the data model at all here, just the ui and a handfull of
backend functions.

i'm trying not to rock the boat too much here, so all my new/remodled subs
are going into a temp module tnmc::pics::new, i'll have to find final
resting places for everything later. we'll see more changes in the months
to come....


git-svn-id: file:///tmp/svnrepo/trunk/tnmc@502 cf5a613f-8eca-db11-8eb1-000ea64e904d